When running a query, statistical data can be collected by using
a Categories or Histogram filter. Previously these filters assumed that
the data was continuous, and stored it in a series of histograms
using a logarithmic scale for the buckets.

Now the filters are configurable, and can (currently) use one of three
bucketing schemes:

- `log`: Use increasingly larger buckets for values further from zero.
- `logQuantile`: As `log`, but use regularly-sized buckets within each
  order of magnitude.
- `categorical`: Store every unique value in its own bucket.

The default is `logQuantile`.

Merge branch 'feature-hist-config' into akka-workers

+++ src/spatialcubeservice/src/main/java/org/vpac/web/model/response/TableRow.java
@@ -5,15 +5,15 @@ import javax.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement(name = "TableRow")
 public class TableRow {
-	private Integer id;
+	private double id;
 	private double area;
 
-	public Integer getId() {
+	public double getId() {
 		return id;
 	}
 
 	@XmlAttribute
-	public void setId(Integer id) {
+	public void setId(double id) {
 		this.id = id;
 	}
 
@@ -29,7 +29,7 @@ public class TableRow {
 	public TableRow() {
 	}
 
-	public TableRow(Integer id, double area) {
+	public TableRow(double id, double area) {
 		this.id = id;
 		this.area = area;
 	}
